Distance is now float64

+++ src/distance/ros-1-wrapper.py
@@ -27,7 +27,7 @@ if __name__ == "__main__":
     rospy.init_node("distance")
     use_twist = rospy.get_param("use_twist",'False')
     d = distance.Distance(use_twist)
-    pub = rospy.Publisher("/distance", Int32, queue_size=0)
+    pub = rospy.Publisher("/distance", Float64, queue_size=0) 
     rospy.Subscriber("/odom", Odom, callbackOdoom)
     rospy.Subscriber("/cmd_vel",Twist , callbackTwist)
     s = rospy.Service('set_dist', SetDist, handle_add_two_ints)
